Add preliminary TypeReplacer support
Adding the TypeReplacer class with supporting OperationReplacer class extending
the OperationCloner class. Cleaned up creation of Pointer and Struct types in
the Base extension. With this commit, TypeReplacer is present but has not yet
been used or tested. But the corresponding changes needed in all other classes
have been made and existing tests function correctly.

Signed-off-by: mstoodle <mstoodle@ca.ibm.com>