Initial VM extension implementation with tests
Adds support for a VM extension with BytecodeBuilder, VirtualMachineState,
VirtualMachineRegister, VirtualMachineRegisterInStruct, and
VirtualMachineOperandStack. Tests are added for virtual machine registers
and for the operand stack, but these are just the code samples from
JitBuilder ported to use the VM extension and JitBuilder 2. These tests
should really be greatly expanded to become better unit tests.

Signed-off-by: mstoodle <mstoodle@ca.ibm.com>